This commit DRAMATICALLY increases server security, at the slight cost of a tiny bit developer sanity.
We no longer store a version of a user's private key in mongo at all.
We not ONLY store an encrypted version of that key.

We give the client a passkey (A hash based on their password) and use that to unlock the key when needed.

This means that the server can't post on your behalf without you logged in.

Further, it means that if the server is hacked ($DEITY forbid), the badguys won't get your privkey.

W00t.