It's always difficult to reverse a design decision you made earlier, particularly one you thought was clever at the time..
In this case, I'm removing the cookie which stores the whole UserObj..
It felt like a really cool idea at the time- I could get any info, no matter where I was..
But it was getting bigger and bigger, esp as we started adding additional Keys..
It was sent with every req (bad), and worse, Safari only allows 4Kb/site.

Safari's decision was the right one- It forced me to stop being a doofus and just store an ID, like everyone else.

Anyway, this could be better optimized (Store in mem!!), but it works for now :/