Webtav will now cache pages for users who aren't logged in. This is likely to represent a large portion of traffic - People who are casually browsing, bots, spiders, and the like.
In Reddit, about 50% of people are logged in.
(Per Jeremy Edberg - http://www.quora.com/What-percentage-of-Reddits-traffic-is-by-registered-users)

The tricky part was doing this while preserving unique XSRF tokens for each user.
We do NOT want to cache the whole page, since doing so would then either
a)    Not use XSRF tokens for unlogged in users.
b)    Share/cache XSRF tokens for unlogged in users.
Both of these scenerios are dangerous.

Instead, we're using nginx SSI to pull and cache the page (by sessionid, arguments, etc)
Then, we cache it on disk in nginx, and serve it up to the anonymous users.